Successfully completed checkoff 1 and 2
Implemented simple LED toggling from the received char via putty for the first checkoff. Implemented a more complex scheme for requesting an LED char, then determining the desired state based on 0 for off, 1 for on, and 2 for toggle. This code implements a lot of flags to allow for a more robust application.